ITMO_ALGS: Fixed sscanf specifier problem 4 W1T4.

+++ itmo_algs/w1/4/main.h
@@ -2,7 +2,7 @@
 
 typedef struct el 
 {
-    long double value;
+    double value;
     unsigned int idx;
 } element_t;
